+++ apps/include/s_apps.h
