+++ test/testutil.h
@@ -15,7 +15,7 @@
 #include <openssl/err.h>
 #include <openssl/e_os2.h>
 #include <openssl/bn.h>
-#include "../apps/opt.h"
+#include "opt.h"
 
 /*-
  * Simple unit tests should implement setup_tests().
